05-hacka/pyproject.toml
[project]
name = "05-hacka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
]

[dependency-groups]
dev = ["isort>=6.0.1", "ruff>=0.11.13"]

[tool.ruff]
line-length = 150

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 150

[tool.ruff.lint]
ignore = ["E402", "F403", "F405"]

[tool.ruff.lint.isort]
case-sensitive = true
combine-as-imports = true
